Adding a staging config file.  (#32)
* Adds necessary capabilities to scamper binary

We would like to run all M-Lab container processes as a non-root user.
It turns out that the scamper binary requires a good number of
privileged capabilities. This commit adds those capabilities to the
scamper binary so that the container can run as non-root. This same
thing was done for scamper in the M-Lab traceroute-caller sidecar:

https://github.com/m-lab/traceroute-caller/pull/162

* Adds cap_net_raw to the revtrvp binary

It turns out that revtrvp needs to create raw sockets to listen for ICMP
packets.

* Adding a log error when we cannot start the socket to listen for spoofed probes

* Merge branch 'master' of https://github.com/NEU-SNS/revtrvp

* Adding a flag to set whether the node should be able to receive spoofed packets or not

* Merge pull request #9 from m-lab/sandbox-kinkade

Adds necessary capabilities to scamper binary

* removing old root.crt

* increasing sending rate

* Adding new assume symmetric assumtion types

* Adding new assume symmetric assumtion types

* Scamper server TLS authentifaction

* adding config

* Merge branch 'main' into master

* Fiwing Dockerfile

* Merge branch 'master' of https://github.com/NEU-SNS/revtrvp

* staging config file

* plvp.config file is now for statging

* good parameters for mlab configs

Co-Authored-By: Nathan Kinkade <kinkade@measurementlab.net>